remove clear ioapic
Remove call to clear_IO_APIC in early boot

In hyper-v we can expect the IOAPIC to be empty on boot. The call
to clear_IO_APIC is then not necessary or desirable because it can
cause significant boot delays in hyper-v. We were seeing up to 300 ms
delays with this call on 32-core systems.